simplified edit page
